support pal/screenManager
