+++ @types/pal/system-info.d.ts
@@ -5,7 +5,6 @@ declare module 'pal/systemInfo' {
         readonly gl: boolean;
         readonly canvas: boolean;
         readonly imageBitmap: boolean;
-        readonly fullscreen: boolean,
     }
 
     class SystemInfo {
@@ -37,42 +36,6 @@ declare module 'pal/systemInfo' {
         public get browserVersion (): string;
         public get pixelRatio (): number;
         public get supportCapability (): SupportCapability;
-        public get isOnFullScreen (): boolean;
-
-        /**
-         * Asynchronously resize screen to a specified size.
-         * Available on desktop related platform.
-         * @param size Specify the size that the screen need to resize to.
-         * @returns Promise to resize screen.
-         * @todo not implemented yet
-         */
-        public resizeScreen (size: import('cocos/core/math').Size): Promise<void>;
-        /**
-         * Asynchronously request fullscreen
-         * If failed to request fullscreen, another attempt will be made to request fullscreen the next time a user interaction occurs.
-         * @returns Promise to request fullscreen
-         */
-        public requestFullScreen (): Promise<void>;
-        /**
-         * Asynchronously exit fullscreen
-         * @returns Promise to exit fullscreen
-         */
-        public exitFullScreen (): Promise<void>;
-        /**
-         * Get the size of current screen.
-         * On Web platform, this should be the size of game container.
-         */
-        public getScreenSize (): import('cocos/core/math').Size;
-        /**
-         * Get the orientation of current game.
-         * Available on mobile related platform.
-         */
-        public getOrientation (): import('pal/system-info/enum-type').Orientation;
-        /**
-         * Get the SafeAreaEdge based on the screen coordinate system.
-         * @return {SafeAreaEdge} An interface displaying the distance of the sides 'top', 'bottom', 'left' and 'right'.
-         */
-        public getSafeAreaEdge (): SafeAreaEdge;
         public getBatteryLevel (): number;
 
         public triggerGC (): void;
@@ -85,26 +48,13 @@ declare module 'pal/systemInfo' {
         public onHide (cb: () => void);
         public onShow (cb: () => void);
         public onClose (cb: () => void);
-        public onFullscreenChange (cb: () => void);
-        public onScreenResize (cb: () => void);
-        public onOrientationChange (cb: () => void);
 
         public offHide (cb?: () => void);
         public offShow (cb?: () => void);
         public offClose (cb?: () => void);
-        public offFullscreenChange (cb?: () => void);
-        public offScreenResize (cb?: () => void);
-        public offOrientationChange (cb?: () => void);
 
         // TODO: support onError
     }
 
     export const systemInfo: SystemInfo;
-
-    export interface SafeAreaEdge {
-        top: number;
-        bottom: number;
-        left: number;
-        right: number;
-    }
 }
