+++ pal/minigame/runtime.ts
@@ -1,7 +1,7 @@
 import { COCOSPLAY, HUAWEI, LINKSURE, OPPO, QTT, VIVO } from 'internal:constants';
 import { SystemInfo, IMiniGame } from 'pal/minigame';
 
-import { Orientation } from '../system-info/enum-type/orientation';
+import { Orientation } from '../screen-manager/enum-type';
 import { cloneObject, createInnerAudioContextPolyfill } from '../utils';
 
 declare let ral: any;
