+++ pal/screen-manager/enum-type/orientation.ts
