+++ pal/screen-manager/enum-type/screen-event.ts
@@ -1,0 +1,5 @@
+export enum ScreenEvent {
+    SCREEN_RESIZE = 'screen_resize',
+    ORIENTATION_CHANGE = 'orientation_change',
+    FULLSCREEN_CHANGE = 'fullscreen_change',
+}
