+++ pal/screen-manager/web/screen-manager.ts
@@ -1,4 +1,7 @@
+import { SafeAreaEdge } from 'pal/screenManager';
+import { EventTarget } from '../../../cocos/core';
 import { Size } from '../../../cocos/core/math';
+import { ScreenEvent, Orientation } from '../enum-type';
 
 interface IScreenFunctionName {
     requestFullscreen: string,
@@ -9,7 +12,8 @@ interface IScreenFunctionName {
     fullscreenerror: string,
 }
 
-class ScreenUtils {
+class ScreenManager {
+    private _eventTarget: EventTarget = new EventTarget();
     private _gameFrame?: HTMLDivElement;
     private _supportFullScreen = false;
     private _touchEventName: string;
@@ -60,11 +64,6 @@ class ScreenUtils {
         ],
     ];
 
-    /**
-     * Setting the fullscreen change callback
-     */
-    public fullScreenChangeCallback?: () => void;
-
     constructor () {
         // TODO: need to access frame from 'pal/launcher' module
         this._gameFrame = document.getElementById('GameDiv') as HTMLDivElement;
@@ -84,34 +83,58 @@ class ScreenUtils {
 
         this._supportFullScreen = (this._fn.requestFullscreen !== undefined);
         this._touchEventName = ('ontouchstart' in window) ? 'touchstart' : 'mousedown';
+        this._registerEvent();
+    }
+
+    private _registerEvent () {
         document.addEventListener(this._fn.fullscreenerror, () => {
             this._onFullscreenError?.();
         });
         document.addEventListener(this._fn.fullscreenchange, () => {
             this._onFullscreenChange?.();
         });
+
+        window.addEventListener('resize', () => {
+            this._eventTarget.emit(ScreenEvent.SCREEN_RESIZE);
+        });
+        window.addEventListener('orientationchange', () => {
+            this._eventTarget.emit(ScreenEvent.ORIENTATION_CHANGE);
+        });
         document.addEventListener(this._fn.fullscreenchange, () => {
-            this.fullScreenChangeCallback?.();
+            this._eventTarget.emit(ScreenEvent.FULLSCREEN_CHANGE);
         });
     }
 
-    public get supportsFullScreen () {
+    public get supportFullScreen (): boolean {
         return this._supportFullScreen;
     }
-
-    public get isOnFullscreen () {
+    public get isOnFullScreen (): boolean {
         if (!this._supportFullScreen) {
             return false;
         }
         return !!document[this._fn.fullscreenElement];
     }
-
-    public getScreenSize (): Size {
+    public get screenSize (): Size {
         if (this._gameFrame) {
             return new Size(this._gameFrame.clientWidth, this._gameFrame.clientHeight);
         } else {
             return new Size(window.innerWidth, window.innerHeight);
         }
+    }
+    public get orientation (): Orientation {
+        // TODO
+        throw new Error('Method not implemented.');
+    }
+    public get safeAreaEdge (): SafeAreaEdge {
+        return {
+            top: 0,
+            bottom: 0,
+            left: 0,
+            right: 0,
+        };
+    }
+    public resizeScreen (size: Size): Promise<void> {
+        throw new Error('Method not implemented.');
     }
 
     private _doRequestFullScreen (): Promise<void> {
@@ -135,7 +158,6 @@ class ScreenUtils {
             }
         });
     }
-
     public requestFullScreen (): Promise<void> {
         return new Promise((resolve, reject) => {
             this._doRequestFullScreen().then(() => {
@@ -153,7 +175,6 @@ class ScreenUtils {
             });
         });
     }
-
     public exitFullScreen (): Promise<void> {
         return new Promise((resolve, reject) => {
             const requestPromise = document[this._fn.exitFullscreen]();
@@ -164,6 +185,25 @@ class ScreenUtils {
             resolve();
         });
     }
+
+    public onFullScreenChange (cb: () => void) {
+        this._eventTarget.on(ScreenEvent.FULLSCREEN_CHANGE, cb);
+    }
+    public onScreenResize (cb: () => void) {
+        this._eventTarget.on(ScreenEvent.SCREEN_RESIZE, cb);
+    }
+    public onOrientationChange (cb: () => void) {
+        this._eventTarget.on(ScreenEvent.ORIENTATION_CHANGE, cb);
+    }
+    public offFullScreenChange (cb?: () => void) {
+        this._eventTarget.off(ScreenEvent.FULLSCREEN_CHANGE, cb);
+    }
+    public offScreenResize (cb?: () => void) {
+        this._eventTarget.off(ScreenEvent.SCREEN_RESIZE, cb);
+    }
+    public offOrientationChange (cb?: () => void) {
+        this._eventTarget.off(ScreenEvent.ORIENTATION_CHANGE, cb);
+    }
 }
 
-export const screenUtils = new ScreenUtils();
+export const screenManager = new ScreenManager();
