+++ pal/system-info/enum-type/event.ts
@@ -2,7 +2,4 @@ export enum AppEvent {
     HIDE = 'hide',
     SHOW = 'show',
     CLOSE = 'close',
-    RESIZE = 'resize',
-    ORIENTATION_CHANGE = 'orientation_change',
-    FULLSCREEN_CHANGE = 'fullscreen_change',
 }
