+++ pal/system-info/enum-type/index.ts
@@ -3,5 +3,4 @@ export * from './event';
 export * from './language';
 export * from './network-type';
 export * from './operating-system';
-export * from './orientation';
 export * from './platform';
