+++ tsconfig.json
@@ -33,6 +33,7 @@
 
         // pal
         "./@types/pal/system-info",
+        "./@types/pal/screen-manager",
         "./@types/pal/minigame",
         "./@types/pal/audio",
         "./@types/pal/input"
